# tests/contracts/snapshots/openapi.json
{
  "components": {
    "schemas": {
      "DeletionPage": {
        "additionalProperties": false,
        "description": "Stable ordered tombstones and the next cursor when another page exists.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeletionTombstoneView"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "trace_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "items",
          "next_cursor",
          "trace_id"
        ],
        "title": "DeletionPage",
        "type": "object"
      },
...
      "FeedbackReceipt": {
        "additionalProperties": false,
        "description": "Durable feedback result and resulting transparent lifecycle state.",
        "properties": {
          "corrected_memory_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Corrected Memory Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "feedback_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Feedback Id",
            "type": "string"
          },
          "feedback_type": {
            "$ref": "#/components/schemas/FeedbackType"
          },
          "memory_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "resulting_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryState"
              },
              {
                "type": "null"
              }
            ]
          },
          "resulting_strength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resulting Strength"
          },
          "trace_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "feedback_id",
          "feedback_type",
          "memory_id",
          "corrected_memory_id",
          "resulting_state",
          "resulting_strength",
          "created_at",
          "trace_id"
        ],
        "title": "FeedbackReceipt",
        "type": "object"
      },
      "FeedbackRequest": {
        "additionalProperties": false,
        "description": "Record useful, wrong, missing, or corrected recall feedback.",
        "properties": {
          "correction_summary": {
            "anyOf": [
              {
                "maxLength": 2048,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correction Summary"
          },
          "feedback_type": {
            "$ref": "#/components/schemas/FeedbackType"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "memory_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "recall_trace_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recall Trace Id"
          },
          "tenant_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "feedback_type"
        ],
        "title": "FeedbackRequest",
        "type": "object"
      },
...
      "ForgetReceipt": {
        "additionalProperties": false,
        "description": "Deletion result returned by one command or status lookup.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_code": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "propagation_state": {
            "$ref": "#/components/schemas/DeletionPropagationState"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "target_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "$ref": "#/components/schemas/ForgetTargetType"
          },
          "tombstone_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Tombstone Id",
            "type": "string"
          },
          "trace_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "tombstone_id",
          "target_type",
          "target_id",
          "propagation_state",
          "requested_at",
          "completed_at",
          "error_code",
          "trace_id"
        ],
        "title": "ForgetReceipt",
        "type": "object"
      },
      "ForgetRequest": {
        "additionalProperties": false,
        "description": "Explicitly delete one exact memory or its complete source observation.",
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "target_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Target Id",
            "type": "string"
          },
          "target_type": {
            "$ref": "#/components/schemas/ForgetTargetType"
          },
          "tenant_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "target_type",
          "target_id"
        ],
        "title": "ForgetRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ObservationProcessingJobView": {
        "additionalProperties": false,
        "description": "Public state of one durable observation processing job.",
        "properties": {
          "attempt": {
            "minimum": 0.0,
            "title": "Attempt",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error_code": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "job_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Job Id",
            "type": "string"
          },
          "observation_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Observation Id",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/JobState"
          },
          "trace_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Trace Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "observation_id",
          "state",
          "attempt",
          "error_code",
          "created_at",
          "updated_at",
          "trace_id"
        ],
        "title": "ObservationProcessingJobView",
        "type": "object"
      },
...
      "RememberRequest": {
        "additionalProperties": false,
        "description": "Explicitly retain user- or agent-supplied content.",
        "properties": {
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ended At"
          },
          "evidence_ids": {
            "default": [],
            "items": {
              "maxLength": 255,
              "minLength": 1,
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "memory_type": {
            "$ref": "#/components/schemas/MemoryType"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "summary": {
            "maxLength": 2048,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "tenant_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Tenant Id",
            "type": "string"
          }
        },
        "required": [
          "tenant_id",
          "summary",
          "memory_type",
          "occurred_at"
        ],
        "title": "RememberRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "MindBridge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/deletions": {
      "get": {
        "operationId": "listDeletions",
        "parameters": [
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Tenant Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletionPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Deletions"
      }
    },
    "/v1/deletions/{tombstone_id}": {
      "get": {
        "operationId": "getForgetStatus",
        "parameters": [
          {
            "in": "path",
            "name": "tombstone_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Tombstone Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Forget Status"
      }
    },
    "/v1/feedback": {
      "post": {
        "operationId": "recordFeedback",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Record Feedback"
      }
    },
    "/v1/forget": {
      "post": {
        "operationId": "forget",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Forget"
      }
    },
    "/v1/jobs/{job_id}": {
      "get": {
        "operationId": "getObservationJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationProcessingJobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Observation Job"
      }
    },
    "/v1/memories": {
      "post": {
        "operationId": "remember",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RememberRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Remember"
      }
    },
    "/v1/memories/{memory_id}": {
      "get": {
        "operationId": "getMemory",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "tenant_id",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Tenant Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Memory"
      }
    },
    "/v1/observations": {
      "post": {
        "operationId": "observe",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObserveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObservationReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Observe"
      }
    },
    "/v1/recall": {
      "post": {
        "operationId": "recall",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recall"
      }
    }
  }
}